🚧 chore(payments.test.js): add initial test for default root route
This commit adds an initial test for the default root route in the `payments.test.js` file. The test uses the `tap` testing framework and the `build` helper function to create an instance of the app. It then makes a request to the `/payments/tezos/payment-intent` endpoint and logs the response.