🔧 chore(migrations): add down.sql and up.sql files for updating user tables and certificate info
The down.sql file contains SQL statements for dropping constraints, altering columns, and adding columns to the "public"."users" and "public"."tezos" tables.

The up.sql file contains SQL statements for dropping a column from the "public"."users" table, adding a foreign key constraint to the "public"."tezos" table, and adding columns to the "public"."user_course" table.

These migrations are necessary to update the user tables and certificate information in the database.