📦 chore(tezosGenerate.vue): add new component for generating Tezos payments
This commit adds a new component called `tezosGenerate.vue` to the `ui/components/Payments` directory. The component is responsible for generating Tezos payments. It includes a template with a single `<div>` element and a script section.

The script section imports necessary dependencies such as `mapGetters` from Vuex and `CONTRACT_ADDRESS` from the `constants` file. It defines the component's props, data properties, and computed properties. The `created` lifecycle hook calls the `createPaymentIntent` method.

The `createPaymentIntent` method is responsible for creating a payment intent for Tezos payments. It checks if the user has Tezos information and if the wallet is connected. If not, it logs a message and returns. If the wallet is not connected, it dispatches the `tezosWallet/autoLogin` action and recursively calls itself. If the wallet is connected, it makes a POST request to `/payments/tezos/payment-intent` to create the payment intent. It sets the `tezosPrice` property based on the response and logs some information.

Note: Some lines of code are commented out and console logs are present for debugging purposes.