Fix connections and payment intents (#165)
* 🐛 fix(PxPayments.vue): set clientSecret in elementsOptions to fix issue with payment processing
✨ feat(nuxt.config.js): add 'stripe-element-payment' to the list of transpiled modules to enable usage of StripeElementPayment component
✨ feat(stripe/index.vue): add new page for Stripe payment integration and implement payment functionality using StripeElementPayment component
✨ feat(vue-stripe.js): add 'StripeElementPayment' component to the list of globally registered components for Vue Stripe plugin

* 🔧 chore(config.yaml): remove unused handler_webhook_baseurl property from actions configuration

* 🐛 fix(schemas.js): add 'id' property to the required properties of the 'paymentIntent' object in the schema to ensure it is present
🐛 fix(Stripe.js): change 'payment_method_types' to 'automatic_payment_methods' and set 'enabled' to true to enable automatic payment methods
✨ feat(Stripe.js): add console.log statement to log paymentIntent details for debugging purposes

* 🐛 fix(stripe/index.vue): add conditional rendering for stripe-element-payment component to prevent rendering when clientSecret is not available
🐛 fix(stripe/index.vue): force update the component after updating clientSecret to trigger re-rendering

* ✨ feat(stripe.vue): add Stripe payment component to handle payments using Stripe API

This commit adds a new component called `stripe.vue` to handle payments using the Stripe API. The component includes a template with a Stripe payment element and a checkout button. It also includes a script section with props for price and email, data properties for Stripe publishable key, elements options, and confirm parameters. The component has a created hook that generates a payment intent using the Stripe API and updates the client secret. It also includes a pay method that submits the payment using the Stripe payment element.

* 🆕 feat(stripeGenerate.vue): add new component for generating Stripe payment

This commit adds a new component called `stripeGenerate.vue` which is responsible for generating a Stripe payment. The component includes a template with a Stripe payment element and a button for checkout. It also includes a script section with props for price and email, data for Stripe publishable key and options for appearance and confirm parameters. The component is created on mount and includes a method for generating the payment intent and a method for submitting the payment.

* 🔥 chore(stripe.vue): remove unused stripe component

The stripe.vue component was deleted as it was no longer being used in the project. This commit removes the file from the repository.

* 🔍 chore(stripe/index.vue): add console log to debug route in payment flow

* 📦 chore(tezosGenerate.vue): add new component for generating Tezos payments

This commit adds a new component called `tezosGenerate.vue` to the `ui/components/Payments` directory. The component is responsible for generating Tezos payments. It includes a template with a single `<div>` element and a script section.

The script section imports necessary dependencies such as `mapGetters` from Vuex and `CONTRACT_ADDRESS` from the `constants` file. It defines the component's props, data properties, and computed properties. The `created` lifecycle hook calls the `createPaymentIntent` method.

The `createPaymentIntent` method is responsible for creating a payment intent for Tezos payments. It checks if the user has Tezos information and if the wallet is connected. If not, it logs a message and returns. If the wallet is not connected, it dispatches the `tezosWallet/autoLogin` action and recursively calls itself. If the wallet is connected, it makes a POST request to `/payments/tezos/payment-intent` to create the payment intent. It sets the `tezosPrice` property based on the response and logs some information.

Note: Some lines of code are commented out and console logs are present for debugging purposes.

* 🎨 style(tezosGenerate.vue): improve button styling and add Tezos icon and price label
🔧 refactor(tezosGenerate.vue): change created lifecycle hook to mounted for better code organization
🔧 refactor(tezosGenerate.vue): remove commented out code and console logs
✨ feat(tezosGenerate.vue): add pay method to handle payment functionality for non-blockchain users

* 🎨 style(tezosGenerate.vue): improve readability by adding line breaks and indentation
🐛 fix(tezosGenerate.vue): update condition to check for existence of tezos_info property in $auth.user object using optional chaining operator to prevent error when property is undefined

* 🔀 refactor(_courseId.vue): replace Tezos payment button with a component for generating Tezos payments
🔀 refactor(_courseId.vue): remove unused code for Tezos payment button

* 🔥 refactor(stripeGenerate.vue): remove console.info statements for payment intent and route

* 🔥 refactor(_courseId.vue): remove unused import of CONTRACT_ADDRESS constant to improve code cleanliness and reduce unused code
🔥 refactor(_courseId.vue): remove unused buyTezos method to improve code cleanliness and reduce unused code

* 🔧 fix(PxPayments.vue): remove unnecessary async/await from mounted method to improve code readability and performance
🔧 fix(PxPayments.vue): remove unused methods createToken, createIntent, and console.log statements
✨ feat(PxPayments.vue): add submitPayment method to handle payment submission

* 🔥 refactor(stripe/index.vue): remove console.info statements for payment intent and route