Update resources, dependencies, and UI for chapter video navigation (#323)
* feat: Update resources for chapters in academy_db

This commit updates the resources for several chapters in the academy_db. The resources URLs have been modified to include relevant links for each chapter.

* Feature/marked with dom purify (#321)

* marked library added and babel config changes

* vue domurify installed

* mode from marked plugin, removed

* dom purify with marked tested in the DOM

---------

Co-authored-by: Datzel <davemg71@hotmail.com>

* Revert "Feature/marked with dom purify (#321)" (#322)

This reverts commit d7c65b5e3ca6a7500fd41e6e59d14f3fe8ec68d9.

* feat: Update resources for chapters in academy_db

* feat: Update dependencies and configuration for marked library

This commit updates the dependencies and configuration for the marked library. It includes adding the marked library and making changes to the babel configuration. Additionally, the commit excludes the marked library from the babel loader test in the webpack configuration.

Co-authored-by: Datzel <davemg71@hotmail.com>

* feat: Update chapter video navigation UI

This commit updates the UI for chapter video navigation in the course navigator. It improves the layout and styling of the video player and adds a button to navigate to the next chapter. Additionally, it hides the chapter card in the course navigator when viewing the challenge view.

* feat: Update chapter video navigation UI

This commit updates the UI for chapter video navigation in the course navigator. It improves the layout and styling of the video player and adds a button to navigate to the next chapter. Additionally, it hides the chapter card in the course navigator when viewing the challenge view.

* feat: Add challenge column to public.courses table

This commit adds a new column "challenge" to the "public.courses" table in the database. The column is of type "text" and allows null values. This change is necessary to support the new challenge feature in the course navigator.

Co-authored-by: Datzel <davemg71@hotmail.com>

* feat: Add empty down migration for challenge_ilithya_course

This commit adds an empty down migration for the "challenge_ilithya_course" table in the "academy_db" schema. The down migration is necessary to maintain consistency in the database schema and to support future changes related to the challenge feature in the course navigator.

* feat: Add challenge field to chapter and test components

This commit adds the "challenge" field to the chapter and test components in the course navigator. The "challenge" field is used to display the challenge information for each chapter and test. This change is necessary to support the new challenge feature in the course navigator.

---------

Co-authored-by: Datzel <davemg71@hotmail.com>