Refactor logger, fix error handling, add logging and gifting functionality, update endpoints and queries (#349)
* refactor logger in application

* chore: Add GraphQL query for retrieving modules and chapters in payments.js

* feat: Fix error handling in Login class

The code changes in `Auth.js` fix the error handling in the `Login` class by updating the `logger.error` statements to include the error message. This ensures that the correct error message is thrown when the email or wallet is incorrect during login. The recent user commits and repository commits do not provide any additional relevant information for the commit message.

* feat: Add logging and gifting functionality to Payments class

* feat: Add endpoint for gifting a course

* feat: Add logging for sending welcome, recover password, and thanks for purchase emails

* chore: Remove .envkey file

The `.envkey` file was deleted in this commit.

* feat: Add gift_course mutation to GraphQL schema and actions

The code changes add a new mutation called `gift_course` to the GraphQL schema and actions. This mutation allows users to gift a course to another user by providing the course ID and the user ID.

* feat: Update ACTION_SERVER_ENDPOINT URL in docker-compose.yaml

The code changes in `docker-compose.yaml` update the `ACTION_SERVER_ENDPOINT` URL from `http://host.docker.internal:5000/` to `http://host.docker.internal:3000/`. This change ensures that the Docker container communicates with the correct endpoint for the action server.

* feat: Update GraphQL query to include course information and user questions

The code changes in `certificates.js` update the GraphQL query to include additional information about the course and user questions. This allows for a more comprehensive certificate generation process and better tracking of user progress within the course.

* feat: Add assertUserIsAllowedToGenerateCertificate method and generateCertificate function

The code changes in `Documents.js` add a new method called `assertUserIsAllowedToGenerateCertificate` and update the `generateCertificate` function to call this method. This ensures that the user meets the requirements for generating a certificate, such as being enrolled in the course and completing enough questions.

* feat: Update fakeCertificates data structure

The code changes in `index.vue` remove the hard-coded `fakeCertificates` data and update the data structure to be an empty array. This ensures that the certificates section on the landing page is dynamically populated with real data from the backend.

* feat: Add user_questions table and foreign key constraint

This commit adds the `user_questions` table to the `public_users` database in the `academy_db` schema. It also includes a foreign key constraint on the `user_id` column, referencing the `user_question` table in the `public` schema. This allows for better organization and tracking of user questions within the system.