Add you will lean to Heeey's course (#363)
* feat: Update you_will_learn descriptions for generative art modules

The code changes update the you_will_learn descriptions for the generative art modules in the academy_db. This ensures that the descriptions accurately reflect the content covered in each module.

* feat: Update you_will_learn descriptions and course price

The code changes update the you_will_learn descriptions for the generative art modules in the academy_db. Additionally, the price of the course has been updated to 40.

* chore: Add empty SQL block for down migration

The code changes add an empty SQL block in the down.sql file for the "add_you_will_learn_heeey" migration in the academy_db. This block does nothing and is included for consistency with the up migration.