feat: Add columns for visibility, release date, and discount price to the courses table
This commit adds new columns to the "courses" table in the academy_db. The "visible" column is of type boolean and has a default value of true. The "release_date" column is of type timestamptz and allows null values. The "discount_price" column is of type numeric and allows null values. These columns will be used to track the visibility, release date, and discount price of each course.