Add columns for visibility, release date, and discount price to courses table and update related code (#364)
* chore: Update start script to use pino-pretty for logging

* feat: Add columns for visibility, release date, and discount price to the courses table

This commit adds new columns to the "courses" table in the academy_db. The "visible" column is of type boolean and has a default value of true. The "release_date" column is of type timestamptz and allows null values. The "discount_price" column is of type numeric and allows null values. These columns will be used to track the visibility, release date, and discount price of each course.

* feat: Add discount price to course GraphQL query

The code changes in `payments.js` add the `discount_price` field to the GraphQL query for courses. This allows retrieving the discount price of each course along with other course details.

* feat: Update course price logic in Payments.js

Refactor the code in Payments.js to update the logic for retrieving the price of a course. The changes include adding support for discount prices and updating the return value to include the correct price based on whether a discount price is available or not.

* feat: Add support for filtering visible courses in GraphQL query

The code changes in `index.vue` modify the GraphQL query for retrieving courses to include a filter for visible courses only. This ensures that only courses marked as visible are returned in the query results.

* feat: Refactor GraphQL query for retrieving user courses

The code changes in `index.vue` modify the GraphQL query for retrieving user courses. The changes include updating the query structure to improve readability and maintainability.

* feat: Update course price display in buyCourse page

Refactor the code in `_courseId.vue` to update the course price display in the buyCourse page. The changes include showing the discount price if available, and falling back to the regular price if no discount is applied. This improves the user experience by providing accurate and up-to-date pricing information.

* refactor: Update course price display in buyCourse page

* feat: Update course price display in buyCourse page

Refactor the code in `_courseId.vue` to update the course price display in the buyCourse page. The changes include showing the discount price if available, and falling back to the regular price if no discount is applied. This improves the user experience by providing accurate and up-to-date pricing information.