Refactor dashboard page to update sales information display (#404)
* Add credit card and tezos sales information to dashboard page

* Add total volume credit card and tezos sales information to dashboard page

* Refactor dashboard page to remove console.info statements

* rename path to show dashboard

* Refactor dashboard page to show stats by teacher

* Refactor dashboard page to show aggregated statistics by teacher

* Refactor dashboard page to remove console.info statements

* Refactor dashboard page to fix user_course null reference error

* Refactor dashboard page to update sales information display

* Refactor dashboard page to update sales information display

* Refactor dashboard page to remove console.info statements and update sales information display

* feat: Update GraphQL query for retrieving user courses

* feat: Update GraphQL query for retrieving user courses

* feat: Refactor GraphQL query for retrieving user courses

This commit refactors the GraphQL query in the `index.vue` file to improve the structure and readability. The changes aim to enhance maintainability and make it easier to understand and modify the query in the future.

* feat: Refactor GraphQL query for retrieving user courses

* feat: Refactor dashboard page to improve sales information display

The code changes in this commit refactor the dashboard page to enhance the display of sales information. This includes removing unnecessary console.info statements, fixing a null reference error, and updating the sales information display. The changes aim to improve the user experience and provide more accurate and relevant sales data on the dashboard.