Update smart contract functionality and fix bugs (#409)
* migrate to new version of smartpy (wip)

* example of new syntax

* migrations

* migrate admin interface

* add better logs and update tests

* refactor types, migrate crud, and add user

* feat: Add blind gallery address to smart contract

This commit adds the blind gallery address to the smart contract in `academy2.py`. The blind gallery address is set to `tz1UdddbVe3icmr5LRP1monxfR13ChsidcoX`. This change is necessary to enable functionality related to the blind gallery in the smart contract.

* update tests and bug fixes

* feat: Update course payment logic and add teacher cut distribution

This commit updates the payment logic in the `academy2.py` smart contract. It ensures that a user can only buy a course if they have not already paid for it and if the course is not already active. Additionally, it introduces a distribution of the teacher's cut from the course price. The teacher's cut is calculated based on the percentage specified in the course details and is sent to the teacher's address. The remaining amount is sent to the blind gallery address. These changes improve the functionality and fairness of the payment system in the smart contract.

* initialize courses

* update tezos constants

* update tezos constants

* fix typo

* update constant

* migrate old sc

* feat: Add onchainId parameter to tezosVerify function