Changes in payment intents (#413)
* feat: Add payment constants module

This commit adds a new module `payments.js` to the `api/constants` directory. The module exports a constant object `TRANSACTION_TYPES` which contains different types of transactions such as `STRIPE`, `TEZOS`, and `GIFT`. This addition enhances the codebase by providing a centralized location for payment-related constants.

* feat: Refactor payment GraphQL queries and mutations

This commit refactors the payment GraphQL queries and mutations in the `payments.js` file. It updates the `GET_PAYMENT_INTENT_INFO` query to use the `payments_by_pk` field instead of the `payments` field. It also updates the `CREATE_PAYMENT_INTENT` mutation to use the `insert_payments_one` field instead of the `insert_payments_one` field. These changes improve the clarity and consistency of the codebase.

* feat: Refactor payment GraphQL queries and mutations

This commit refactors the payment GraphQL queries and mutations in the `Payments.js` file. It updates the `GET_PAYMENT_INTENT_INFO` query to use the `payments_by_pk` field instead of the `payments` field. It also updates the `CREATE_PAYMENT_INTENT` mutation to use the `insert_payments_one` field instead of the `insert_payments_one` field. These changes improve the clarity and consistency of the codebase.

* Refactor payment GraphQL queries and mutations

* feat: Refactor payment GraphQL queries and mutations

This commit refactors the payment GraphQL queries and mutations in the `Payments.js` file. It updates the `GET_PAYMENT_INTENT_INFO` query to use the `payments_by_pk` field instead of the `payments` field. It also updates the `CREATE_PAYMENT_INTENT` mutation to use the `insert_payments_one` field instead of the `insert_payments_one` field. These changes improve the clarity and consistency of the codebase.

* chore: Update npm dependencies for stripe and pino-pretty

* Update transaction types