Refactor code and update dependencies (#426)
* refactor: Remove unnecessary contract variable assignment

* refactor: Remove unnecessary console.info statement in BecomeAnEducatorForm.vue

* chore: Remove unnecessary console.info statement in tezosGenerate.vue

* refactor: Update courseId prop type in openModalButton.vue

The code changes in the `openModalButton.vue` file update the `courseId` prop type from `Boolean` to `String`. This change ensures that the `courseId` prop can accept string values, providing more flexibility in its usage.

* feat: Update mintButton.vue to use soulBoundTokenId for conditional rendering

The code changes in the `mintButton.vue` file update the conditional rendering logic to use the `soulBoundTokenId` property instead of `hash`. This change ensures that the correct button is displayed based on the presence of `soulBoundTokenId`. Additionally, the code adds a loading message when claiming the certificate.

* refactor: Update logger.info statement in tzip.js

Remove unnecessary path parameter from the logger.info statement in the tzip.js file. The path parameter is not needed and can be removed to simplify the code.

* refactor: Update dependency import in sbt.js

Update the import statement for the `char2Bytes` function in the `sbt.js` file to use the `@taquito/utils` package instead of the deprecated `@taquito/tzip16` package. This change ensures that the code uses the latest and recommended package for the desired functionality.

* feat: Update mintButton.vue to use soulBoundTokenId for conditional rendering

* refactor: Update mintButton.vue to use soulBoundTokenId for conditional rendering

Refactor the `mintButton.vue` file to improve the conditional rendering logic by using the `soulBoundTokenId` property instead of `hash`. This change ensures that the correct button is displayed based on the presence of `soulBoundTokenId`. Additionally, a loading message is added when claiming the certificate.