feat: Add customer_id field to user table and sku field to courses table
This commit adds the `customer_id` field to the user table and the `sku` field to the courses table. The `customer_id` field will be used to store the Stripe customer ID associated with each user, while the `sku` field will be used to keep track of the product in Stripe. These changes enable better integration with the Stripe service and improve the functionality of the application.