feat: Add Stripe customer registration during login
This commit adds the functionality to register a customer in Stripe during the login process. If the user does not have a Stripe customer ID, it will be created and associated with the user. This integration with Stripe improves the application's ability to manage customer information and enables seamless payment processing.