feat: Add IP address tracking to login process
This commit adds the functionality to track the IP address of the user during the login process. The `getIp` method is introduced to retrieve the user's IP address using an external API. The IP address is then included in the login request payload, enhancing the application's ability to track and analyze user login activity.