refactor: Add country field to user queries and mutations
This commit adds the "country" field to the user queries and mutations in the GraphQL schema. The "country" field is now included in the "users_by_pk" query, the "users" query, and the "update_users_by_pk" mutation. This change allows for retrieving and updating the country information for users.