feat: Update Tezos RPC endpoints to use new URLs
This commit updates the Tezos RPC endpoints in the codebase to use the new URLs provided. The `mainnet` endpoint is now set to 'https://rpc.tzkt.io/mainnet' and the `ghostnet` endpoint is set to 'https://rpc.ghostnet.teztnets.com/'. This change ensures that the application is using the correct RPC endpoints for Tezos network communication.